Added logging when the mmg callback is not json. Refactored the tests to match structure of notifications.
